b4

+++ scripts/cpython-build-host-deps.sh
@@ -9,18 +9,26 @@ echo "
 # python3 setup.py install --single-version-externally-managed --root=/
 SPY="$HPY setup.py install --single-version-externally-managed --root=/"
 
+
 # just in case
 $PIP install pip --upgrade
 
-# to remove ctypes deps
-$PIP install setuptools --upgrade
 
-HPFX=./devices/x86_64/usr/lib/python3.${PYMINOR}
-rm $HPFX/ensurepip/_bundled/setuptools-*-py3-none-any.whl
-mv $HPFX/site-packages/setuptool* $HPFX/
-mv $HPFX/site-packages/_distutils* $HPFX/
-mv $HPFX/site-packages/pkg_resources $HPFX/
+if [ -d $HPFX/site-packages/setuptools ]
+then
+    echo "
+    * setuptools/_distutils/pkg_resources already migrated
+"
+else
+    # to remove ctypes deps
+    $PIP install setuptools --upgrade
 
+    HPFX=./devices/x86_64/usr/lib/python3.${PYMINOR}
+    rm $HPFX/ensurepip/_bundled/setuptools-*-py3-none-any.whl
+    mv $HPFX/site-packages/setuptools* $HPFX/
+    mv $HPFX/site-packages/_distutils* $HPFX/
+    mv $HPFX/site-packages/pkg_resources* $HPFX/
+fi
 
 
 export CC=clang
